+++ engine/src/main/java/net/jqwik/engine/execution/lifecycle/ExtendedPropertyExecutionResult.java
@@ -15,4 +15,6 @@ public interface ExtendedPropertyExecutionResult extends PropertyExecutionResult
 
 	GenerationMode generation();
 
+	EdgeCasesMode edgeCases();
+
 }